+++ gears/processors/__init__.py
@@ -1,1 +1,2 @@
+from .base import BaseProcessor
 from .directives import DirectivesProcessor