+++ gears/processors/directives.py
@@ -14,6 +14,7 @@ class DirectivesProcessor(BaseProcessor):
             'require': self.process_require_directive,
             'require_directory': self.process_require_directory_directive,
             'require_self': self.process_require_self_directive,
+            'depend_on': self.process_depend_on_directive,
         }
 
     def __call__(self, asset):
@@ -36,13 +37,16 @@ class DirectivesProcessor(BaseProcessor):
 
     def process_require_directory_directive(self, path):
         path = self.get_relative_path(path, is_directory=True)
-        list = self.asset.attributes.environment.list(path, self.asset.attributes.suffix)
+        list = self.asset.attributes.environment.list(path, self.asset.attributes.mimetype)
         for asset_attributes, absolute_path in sorted(list, key=lambda x: x[0].path):
             self.asset.requirements.add(self.get_asset(asset_attributes, absolute_path))
             self.asset.dependencies.add(os.path.dirname(absolute_path))
 
     def process_require_self_directive(self):
         self.asset.requirements.add(self.asset)
+
+    def process_depend_on_directive(self, path):
+        self.asset.dependencies.add(self.find(path)[1])
 
     def find(self, require_path):
         require_path = self.get_relative_path(require_path)