+++ tests/test_asset_attributes.py
@@ -1,14 +1,24 @@
 from gears.asset_attributes import AssetAttributes
+from gears.compilers import BaseCompiler, CoffeeScriptCompiler, StylusCompiler
 from gears.environment import Environment
 
 from mock import Mock
 from unittest2 import TestCase
 
 
+class TemplateCompiler(BaseCompiler):
+
+    def __call__(self, asset):
+        pass
+
+
 class AssetAttributesTests(TestCase):
 
     def setUp(self):
         self.environment = Environment('assets')
+        self.coffee_script_compiler = CoffeeScriptCompiler.as_handler()
+        self.stylus_compiler = StylusCompiler.as_handler()
+        self.template_compiler = TemplateCompiler.as_handler()
 
     def create_attributes(self, path):
         return AssetAttributes(self.environment, path)
@@ -22,6 +32,7 @@ class AssetAttributesTests(TestCase):
         check('js/readme', [])
         check('js/script.js', ['.js'])
         check('js/script.js.coffee', ['.js', '.coffee'])
+        check('js/script.coffee', ['.coffee'])
         check('js/app.min.js.coffee', ['.min', '.js', '.coffee'])
         check('js/.htaccess', ['.htaccess'])
 
@@ -35,12 +46,14 @@ class AssetAttributesTests(TestCase):
         check('js/.htaccess', 'js/')
         check('js/readme', 'js/readme')
         check('js/app.min.js.coffee', 'js/app')
+        check('js/script.coffee', 'js/script')
 
         self.environment.mimetypes.register('.js', 'application/javascript')
         check('js/app.min.js.coffee', 'js/app.min')
 
     def test_logical_path(self):
         self.environment.mimetypes.register('.js', 'application/javascript')
+        self.environment.compilers.register('.coffee', self.coffee_script_compiler)
 
         def check(path, expected_result):
             logical_path = self.create_attributes(path).logical_path
@@ -48,6 +61,7 @@ class AssetAttributesTests(TestCase):
 
         check('js/script.js', 'js/script.js')
         check('js/script.js.coffee', 'js/script.js')
+        check('js/script.coffee', 'js/script.js')
         check('js/script.min.js', 'js/script.min.js')
         check('js/script.min.js.coffee', 'js/script.min.js')
 
@@ -66,7 +80,7 @@ class AssetAttributesTests(TestCase):
     def test_format_extension(self):
         self.environment.mimetypes.register('.css', 'text/css')
         self.environment.mimetypes.register('.js', 'application/javascript')
-        self.environment.compilers.register('.coffee', Mock())
+        self.environment.compilers.register('.coffee', self.coffee_script_compiler)
 
         def check(path, expected_result):
             format_extension = self.create_attributes(path).format_extension
@@ -77,12 +91,13 @@ class AssetAttributesTests(TestCase):
         check('js/jquery.min.js', '.js')
         check('js/app.js.min.coffee', '.js')
         check('css/style.js.css', '.css')
+        check('js/script.coffee', None)
         check('readme', None)
 
     def test_suffix(self):
         self.environment.mimetypes.register('.css', 'text/css')
         self.environment.mimetypes.register('.js', 'application/javascript')
-        self.environment.compilers.register('.coffee', Mock())
+        self.environment.compilers.register('.coffee', self.coffee_script_compiler)
 
         def check(path, expected_result):
             suffix = self.create_attributes(path).suffix
@@ -90,6 +105,7 @@ class AssetAttributesTests(TestCase):
 
         check('js/script.js', ['.js'])
         check('js/script.js.coffee', ['.js', '.coffee'])
+        check('js/script.coffee', ['.coffee'])
         check('js/script.min.js.coffee', ['.js', '.coffee'])
         check('js/script.js.min.coffee', ['.js', '.min', '.coffee'])
         check('readme', [])
@@ -97,8 +113,8 @@ class AssetAttributesTests(TestCase):
     def test_compiler_extensions(self):
         self.environment.mimetypes.register('.css', 'text/css')
         self.environment.mimetypes.register('.js', 'application/javascript')
-        self.environment.compilers.register('.coffee', Mock())
-        self.environment.compilers.register('.tmpl', Mock())
+        self.environment.compilers.register('.coffee', self.coffee_script_compiler)
+        self.environment.compilers.register('.tmpl', self.template_compiler)
 
         def check(path, expected_result):
             compiler_extensions = self.create_attributes(path).compiler_extensions
@@ -106,31 +122,46 @@ class AssetAttributesTests(TestCase):
 
         check('js/script.js', [])
         check('js/script.js.coffee', ['.coffee'])
+        check('js/script.coffee', ['.coffee'])
         check('js/script.js.coffee.tmpl', ['.coffee', '.tmpl'])
         check('js/hot.coffee.js.tmpl', ['.tmpl'])
 
-    def test_compilers(self):
-        coffee_compiler = Mock()
-        template_compiler = Mock()
+    def test_compiler_format_extension(self):
         self.environment.mimetypes.register('.css', 'text/css')
         self.environment.mimetypes.register('.js', 'application/javascript')
-        self.environment.compilers.register('.coffee', coffee_compiler)
-        self.environment.compilers.register('.tmpl', template_compiler)
+        self.environment.compilers.register('.coffee', self.coffee_script_compiler)
+        self.environment.compilers.register('.styl', self.stylus_compiler)
+        self.environment.compilers.register('.tmpl', self.template_compiler)
+
+        def check(path, expected_result):
+            format_extension = self.create_attributes(path).compiler_format_extension
+            self.assertEqual(format_extension, expected_result)
+
+        check('js/application.coffee', '.js')
+        check('css/application.styl', '.css')
+        check('application.tmpl', None)
+
+    def test_compilers(self):
+        self.environment.mimetypes.register('.css', 'text/css')
+        self.environment.mimetypes.register('.js', 'application/javascript')
+        self.environment.compilers.register('.coffee', self.coffee_script_compiler)
+        self.environment.compilers.register('.tmpl', self.template_compiler)
 
         def check(path, expected_result):
             compilers = self.create_attributes(path).compilers
             self.assertEqual(compilers, expected_result)
 
         check('js/script.js', [])
-        check('js/script.js.coffee', [coffee_compiler])
-        check('js/script.js.coffee.tmpl', [coffee_compiler, template_compiler])
-        check('js/hot.coffee.js.tmpl', [template_compiler])
+        check('js/script.js.coffee', [self.coffee_script_compiler])
+        check('js/script.coffee', [self.coffee_script_compiler])
+        check('js/script.js.coffee.tmpl', [self.coffee_script_compiler, self.template_compiler])
+        check('js/hot.coffee.js.tmpl', [self.template_compiler])
 
     def test_mimetype(self):
         self.environment.mimetypes.register('.css', 'text/css')
         self.environment.mimetypes.register('.js', 'application/javascript')
-        self.environment.compilers.register('.coffee', Mock())
-        self.environment.compilers.register('.styl', Mock())
+        self.environment.compilers.register('.coffee', self.coffee_script_compiler)
+        self.environment.compilers.register('.styl', self.stylus_compiler)
 
         def check(path, expected_result):
             mimetype = self.create_attributes(path).mimetype
@@ -138,6 +169,7 @@ class AssetAttributesTests(TestCase):
 
         check('js/script.js', 'application/javascript')
         check('js/script.js.coffee', 'application/javascript')
+        check('js/script.coffee', 'application/javascript')
         check('css/style.min.css', 'text/css')
         check('readme.txt', 'application/octet-stream')
 
@@ -146,9 +178,9 @@ class AssetAttributesTests(TestCase):
         second_processor = Mock()
         self.environment.mimetypes.register('.css', 'text/css')
         self.environment.mimetypes.register('.js', 'application/javascript')
-        preprocessors = self.environment.preprocessors
-        preprocessors.register('text/css', first_processor)
-        preprocessors.register('text/css', second_processor)
+        self.environment.compilers.register('.styl', self.stylus_compiler)
+        self.environment.preprocessors.register('text/css', first_processor)
+        self.environment.preprocessors.register('text/css', second_processor)
 
         def check(path, expected_result):
             preprocessors = self.create_attributes(path).preprocessors
@@ -157,15 +189,17 @@ class AssetAttributesTests(TestCase):
         check('readme.txt', [])
         check('js/script.js', [])
         check('css/style.css', [first_processor, second_processor])
+        check('css/style.css.styl', [first_processor, second_processor])
+        check('css/style.styl', [first_processor, second_processor])
 
     def test_postprocessors(self):
         first_processor = Mock()
         second_processor = Mock()
         self.environment.mimetypes.register('.css', 'text/css')
         self.environment.mimetypes.register('.js', 'application/javascript')
-        postprocessors = self.environment.postprocessors
-        postprocessors.register('text/css', first_processor)
-        postprocessors.register('text/css', second_processor)
+        self.environment.compilers.register('.styl', self.stylus_compiler)
+        self.environment.postprocessors.register('text/css', first_processor)
+        self.environment.postprocessors.register('text/css', second_processor)
 
         def check(path, expected_result):
             postprocessors = self.create_attributes(path).postprocessors
@@ -174,3 +208,5 @@ class AssetAttributesTests(TestCase):
         check('readme.txt', [])
         check('js/script.js', [])
         check('css/style.css', [first_processor, second_processor])
+        check('css/style.css.styl', [first_processor, second_processor])
+        check('css/style.styl', [first_processor, second_processor])